primex-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ollama==0.1.6
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
httpx>=0.25.2,<0.26.0
